refactor: Relocate movie response DTOs
