+++ data/src/main/java/com/madrid/data/dataSource/remote/RemoteDataSourceImpl.kt
@@ -18,8 +18,8 @@ import com.madrid.data.dataSource.remote.dto.series.SeriesCreditResponse
 import com.madrid.data.dataSource.remote.dto.series.SeriesDetailsResponse
 import com.madrid.data.dataSource.remote.dto.series.SeriesReviewResponse
 import com.madrid.data.dataSource.remote.dto.series.SimilarSeriesResponse
-import com.madrid.data.dataSource.remote.response.movie.NowPlayingMovieResponse
-import com.madrid.data.dataSource.remote.response.movie.UpcomingMoviesResponse
+import com.madrid.data.dataSource.remote.dto.movie.NowPlayingMovieResponse
+import com.madrid.data.dataSource.remote.dto.movie.UpcomingMoviesResponse
 import com.madrid.data.dataSource.remote.response.series.AiringTodayTvShowsResponse
 import com.madrid.data.dataSource.remote.response.series.OnAirTvShowsResponse
 import com.madrid.data.dataSource.remote.response.series.RecommendedSeriesResponse
